requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.47.0
Pillow>=10.3.0
numpy>=1.26.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
opencv-python-headless>=4.9.0
httpx>=0.27.0
einops>=0.7.0
kornia>=0.7.0
timm>=0.9.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
